Add a new target

+++ include/cgpa_calc.hpp
@@ -5,7 +5,7 @@ class cgpa_calc
    public:
       double getCGPA(void);
       double calculateCGPA(void);
-      void setMarks(int maths, int phy, int chem, int eng);
+      void setMarks(double maths, double phy, double chem, double eng);
 
    protected:
       double maths, physics, chemistry, english, cgpa;
